chore: rename lib/functions lib/utils
