+++ workers/fund_public_goods/lib/strategy/utils/assign_weights.py
