+++ workers/fund_public_goods/lib/strategy/utils/generate_queries.py
